CSC486B/DeepLearning_in_Computer_Vision/assignment1/requirements.txt
appnope==0.1.0
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Markdown==2.6.10
matplotlib==2.1.1
numpy==1.13.3
olefile==0.44
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.13
